Backend/requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
pinecone-client
langchain-pinecone
sentence-transformers
langchain-google-genai
tenacity
pydantic
psycopg2-binary
modelcontextprotocol
langchain-openai
langchain-groq
langchain-nvidia-ai-endpoints
mcp
fastapi
uvicorn
python-multipart
redis
sqlglot
tiktoken
openpyxl
langchain-neo4j
neo4j